Add CMake Builds as a GitHub Workflow
Windows doesn't work for some reason, and I think that's out of my control. This uploads to Artifacts, so you have to take a look at the Run to see the files. They also include the cmake build files because I don't want to figure out how to only ship the exec. Not that big of a deal, right?